feat(framework-issues-ui): Refactor AutomatedChecksControl (#1353)
